src/program_of_thought/program_of_thought.py
import logging
import os

import dspy
from dotenv import load_dotenv

logger = logging.getLogger(__name__)


class ProgramOfThought(dspy.Module):
    pass


def program_of_thought(prompt: str) -> str:
    """
    Generates a response using the Program of Thought approach.

    Args:
        prompt (str): The input prompt to process.
    Returns:
        str: The generated response.
    """
    logger.info("Generating response for prompt: %s", prompt)
    load_dotenv()
    llama31_70b = dspy.LM(
        model=os.getenv("TOGETHERAI_API_MODEL"),
        api_key=os.getenv("TOGETHERAI_API_KEY"),
    )
    dspy.settings.configure(
        lm=llama31_70b,
        max_tokens=2048,
        temperature=0.7,
        top_p=0.9,
        stop_sequences=["\n\n"],
    )
    return "Placeholder response for the Program of Thought approach. "
