pyproject.toml
[project]
name = "program-of-thought"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "William C Grisaitis", email = "wgrisaitis@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.27",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "detect-secrets==1.5.0",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
install_types = true
non_interactive = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"

[project.scripts]
program-of-thought = "program_of_thought.cli:main"
